+++ ql/src/java/org/apache/hadoop/hive/ql/udf/UDFToFloat.java
@@ -31,123 +31,130 @@ import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.NullWritable;
 import org.apache.hadoop.io.Text;
 
-
 public class UDFToFloat extends UDF {
 
   private static Log LOG = LogFactory.getLog(UDFToFloat.class.getName());
 
   FloatWritable floatWritable = new FloatWritable();
-  
+
   public UDFToFloat() {
   }
 
   /**
    * Convert from void to a float. This is called for CAST(... AS FLOAT)
-   *
-   * @param i The void value to convert
+   * 
+   * @param i
+   *          The void value to convert
    * @return FloatWritable
    */
-  public FloatWritable evaluate(NullWritable i)  {
+  public FloatWritable evaluate(NullWritable i) {
     return null;
   }
 
   /**
    * Convert from boolean to a float. This is called for CAST(... AS FLOAT)
-   *
-   * @param i The boolean value to convert
+   * 
+   * @param i
+   *          The boolean value to convert
    * @return FloatWritable
    */
-  public FloatWritable evaluate(BooleanWritable i)  {
+  public FloatWritable evaluate(BooleanWritable i) {
     if (i == null) {
       return null;
     } else {
-      floatWritable.set(i.get() ? (float)1.0 : (float)0.0);
+      floatWritable.set(i.get() ? (float) 1.0 : (float) 0.0);
       return floatWritable;
     }
   }
-  
+
   /**
    * Convert from byte to a float. This is called for CAST(... AS FLOAT)
-   *
-   * @param i The byte value to convert
+   * 
+   * @param i
+   *          The byte value to convert
    * @return FloatWritable
    */
-  public FloatWritable evaluate(ByteWritable i)  {
+  public FloatWritable evaluate(ByteWritable i) {
     if (i == null) {
       return null;
     } else {
-      floatWritable.set((float)i.get());
+      floatWritable.set(i.get());
       return floatWritable;
     }
   }
-  
+
   /**
    * Convert from short to a float. This is called for CAST(... AS FLOAT)
-   *
-   * @param i The short value to convert
+   * 
+   * @param i
+   *          The short value to convert
    * @return FloatWritable
    */
-  public FloatWritable evaluate(ShortWritable i)  {
+  public FloatWritable evaluate(ShortWritable i) {
     if (i == null) {
       return null;
     } else {
-      floatWritable.set((float)i.get());
+      floatWritable.set(i.get());
       return floatWritable;
     }
   }
-  
+
   /**
    * Convert from integer to a float. This is called for CAST(... AS FLOAT)
-   *
-   * @param i The integer value to convert
+   * 
+   * @param i
+   *          The integer value to convert
    * @return FloatWritable
    */
-  public FloatWritable evaluate(IntWritable i)  {
+  public FloatWritable evaluate(IntWritable i) {
     if (i == null) {
       return null;
     } else {
-      floatWritable.set((float)i.get());
+      floatWritable.set(i.get());
       return floatWritable;
     }
   }
-  
+
   /**
    * Convert from long to a float. This is called for CAST(... AS FLOAT)
-   *
-   * @param i The long value to convert
+   * 
+   * @param i
+   *          The long value to convert
    * @return FloatWritable
    */
-  public FloatWritable evaluate(LongWritable i)  {
+  public FloatWritable evaluate(LongWritable i) {
     if (i == null) {
       return null;
     } else {
-      floatWritable.set((float)i.get());
+      floatWritable.set(i.get());
       return floatWritable;
     }
   }
 
   /**
    * Convert from double to a float. This is called for CAST(... AS FLOAT)
-   *
-   * @param i The double value to convert
+   * 
+   * @param i
+   *          The double value to convert
    * @return FloatWritable
-   */  
-  public FloatWritable evaluate(DoubleWritable i)  {
+   */
+  public FloatWritable evaluate(DoubleWritable i) {
     if (i == null) {
       return null;
     } else {
-      floatWritable.set((float)i.get());
+      floatWritable.set((float) i.get());
       return floatWritable;
     }
   }
-  
+
   /**
    * Convert from string to a float. This is called for CAST(... AS FLOAT)
-   *
-   * @param i The string value to convert
+   * 
+   * @param i
+   *          The string value to convert
    * @return FloatWritable
    */
-  public FloatWritable evaluate(Text i)  {
+  public FloatWritable evaluate(Text i) {
     if (i == null) {
       return null;
     } else {
@@ -161,5 +168,5 @@ public class UDFToFloat extends UDF {
       }
     }
   }
-  
+
 }
